requirements.txt
fastapi==0.109.0
uvicorn==0.26.0
pydantic==2.5.3
boto3==1.34.25
aiohttp==3.9.1
python-multipart==0.0.6
python-dotenv==1.0.0
Pillow==10.1.0
openai-whisper
yt-dlp
ffmpeg-python==0.2.0
aiofiles==23.2.1 

minio>=7.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5